remove sagemath dependency
