+++ requirements/dev.txt
@@ -1,0 +1,5 @@
+numpy>=1.17
+numba>=0.54.0
+pytest>=6.2.5
+scipy
+sympy
