+++ requirements/requirements.txt
@@ -1,0 +1,2 @@
+numpy>=1.17
+numba>=0.54.0
