+++ setup.cfg
@@ -21,7 +21,7 @@ keywords =
 
 [options]
 packages = agemo
-python_requires = >=3.7
+python_requires = >=3.8
 install_requires = 
     numpy>=1.18
     numba>=0.54.0
@@ -29,8 +29,8 @@ install_requires =
 [options.extras_require]
 tests = 
     pytest>=6.2.5
-    sage-conf
-    sagemath-standard>9.2
+    scipy
+    sympy
 
 [tool:pytest]
 testpaths =
